pyproject.toml
[tool.poetry]
name = "target-non-target-classification"
version = "0.1.0"
description = ""
authors = ["Ryo Shirai <ryo.shirai@linecorp.com>"]
readme = "README.md"

packages = [
    { include = "util" }
]

[tool.poetry.dependencies]
python = "^3.9"
hydra-core = "1.3.2"
werkzeug = "2.3.0"
numpy = "1.26.4"
pandas = "1.5.0"
torch = { url = "https://download.pytorch.org/whl/cu117/torch-2.0.1%2Bcu117-cp39-cp39-linux_x86_64.whl" }
mlflow = "^2.19.0"
scikit-learn = "1.1.2"
matplotlib = "3.6.0"
tqdm = "^4.67.1"
poetry = "^2.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
ruff = "^0.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
check_untyped_defs = true

[tool.ruff]
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".git",
]

line-length = 120
indent-width = 4

[tool.ruff.lint]
select = [
    "C",
    "E",
    "F",
    "W",
]
ignore = [
]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.mccabe]
max-complexity = 10
